% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterviewTPHcompStatistician.R
\docType{package}
\name{InterviewTPHcompStatistician}
\alias{InterviewTPHcompStatistician}
\alias{InterviewTPHcompStatistician-package}
\title{Interview}
\description{
Package containing the second function of the exercise
}
\keyword{internal}
