% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popmean.R
\name{popmean}
\alias{popmean}
\title{National prevalence}
\usage{
popmean(data, intervention, years)
}
\arguments{
\item{data}{Data frame containing the following columns:
zone, population, year, none, IRS, ITN, IRS+ITN}

\item{intervention}{unquoted expression indicating the intervention.}

\item{years}{integer vector of the years the prevalence are calculated.}
}
\value{
a data frame (tibble) with two columns: the year and the national
prevalence for the inputted intervention.
}
\description{
Calculate the national prevalence of an intervention type
for a range of years.
}
\examples{
popmean(dat, intervention = IRS, years = 2020:2022)
popmean(dat, intervention = ITN, years = 2020:2022)
}
